refector : API 경로 수정 - Competition 컨트롤러 통합, 분리, 일부 삭제 #193
