+++ src/main/java/github/com/jbabe/web/controller/competition/CompetitionController.java
@@ -1,5 +1,8 @@
 package github.com.jbabe.web.controller.competition;
 
+import github.com.jbabe.service.competition.CompetitionDetailService;
+import github.com.jbabe.service.competition.CompetitionResultService;
+import github.com.jbabe.service.competition.CompetitionScheduleService;
 import github.com.jbabe.service.competition.CompetitionService;
 import github.com.jbabe.service.exception.BadRequestException;
 import github.com.jbabe.service.userDetails.CustomUserDetails;
@@ -29,9 +32,12 @@ import java.util.Optional;
 public class CompetitionController {
 
     private final CompetitionService competitionService;
+    private final CompetitionDetailService competitionDetailService;
+    private final CompetitionResultService competitionResultService;
+    private final CompetitionScheduleService competitionScheduleService;
 
 
-
+    //목록조회
     @Operation(summary = "대회정보 목록 조회")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "400", description = "BAD_REQUEST",
@@ -55,9 +61,6 @@ public class CompetitionController {
     return new ResponseDto(data);
     }
 
-
-
-
     @Operation(summary = "년도 LIST 조회")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "404", description = "NOT_FOUND",
@@ -65,14 +68,67 @@ public class CompetitionController {
                             examples = @ExampleObject(name = "대회를 찾을 수 없음", value = "{\n  \"code\": 404,\n  \"message\": \"NOT_FOUND\",\n \"detailMessage\": \"대회를 찾을 수 없습니다.\",\n \"request\": null\n}")
                     ))
     })
-    @GetMapping("/find-year-list")
+    @GetMapping("/years")
     public ResponseDto getCompetitionYearList() {
         return new ResponseDto(competitionService.getCompetitionYearList());
     }
 
     @Operation(summary = "종별 LIST 조회")
-    @GetMapping("/find-division-list")
+    @GetMapping("/divisions")
     public ResponseDto getCompetitionDivisionList() {
         return new ResponseDto(competitionService.getCompetitionDivisionList());
     }
+
+
+    //상세조회
+    @Operation(summary = "대회정보 조회")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "404", description = "NOT_FOUND",
+                    content = @Content(mediaType = "application/json",
+                            examples = {
+                                    @ExampleObject(name = "대회를 찾을 수 없음", value = "{\n  \"code\": 404,\n  \"message\": \"NOT_FOUND\",\n \"detailMessage\": \"해당 아이디와 일치하는 대회를 찾을 수 없습니다.\",\n \"request\": \"1\"\n}"),
+                                    @ExampleObject(name = "삭제되거나 숨겨진 대회", value = "{\n  \"code\": 404,\n  \"message\": \"NOT_FOUND\",\n \"detailMessage\": \"대회 조회가 불가능합니다.\",\n \"request\": \"DELETE OR HIDE\"\n}")
+                            }
+                    ))
+    })
+    @GetMapping("/{id}")
+    public ResponseDto getCompetitionDetail(@PathVariable Integer id) {
+        CompetitionDetailResponse data = competitionDetailService.getCompetitionDetail(id);
+        return new ResponseDto(data);
+    }
+
+
+    @Operation(summary = "대회일정 조회")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "404", description = "NOT_FOUND",
+                    content = @Content(mediaType = "application/json",
+                            examples = {
+                                    @ExampleObject(name = "대회를 찾을 수 없음", value = "{\n  \"code\": 404,\n  \"message\": \"NOT_FOUND\",\n \"detailMessage\": \"대회를 찾을 수 없습니다.\",\n \"request\": \"1\"\n}"),
+                                    @ExampleObject(name = "대회일정을 찾을 수 없음", value = "{\n  \"code\": 404,\n  \"message\": \"NOT_FOUND\",\n \"detailMessage\": \"대회일정을 찾을 수 없습니다.\",\n \"request\": \"1\"\n}")
+                            }
+                    ))
+    })
+    @GetMapping("/{id}/schedule")
+    public ResponseDto getCompetitionSchedule (@PathVariable Integer id) {
+        List<GetScheduleResponse> res = competitionScheduleService.getCompetitionSchedule(id);
+        return new ResponseDto(res);
+    }
+
+    @Operation(summary = "대회결과 조회")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "404", description = "NOT_FOUND",
+                    content = @Content(mediaType = "application/json",
+                            examples = {
+                                    @ExampleObject(name = "대회를 찾을 수 없음", value = "{\n  \"code\": 404,\n  \"message\": \"NOT_FOUND\",\n \"detailMessage\": \"대회를 찾을 수 없습니다.\",\n \"request\": \"1\"\n}"),
+                            }
+                    ))
+    })
+    @GetMapping("/{id}/result")
+    public ResponseDto getCompetitionResult (@PathVariable Integer id) {
+        List<GetResultResponse> res = competitionResultService.getCompetitionResult(id);
+        return new ResponseDto(res);
+    }
+
+
+
 }
